requirements.txt
streamlit==1.38.0
langchain==0.2.16
langchain-community==0.2.12
langchain-text-splitters==0.2.2
openai==1.51.0
chromadb==0.5.5
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
